requirements.txt
annotated-types==0.7.0
anyio==4.5.2
ascii==3.6
asttokens==2.4.1
attrs==25.1.0
backcall==0.2.0
base58==2.1.1
beautifulsoup4==4.13.3
blinker==1.8.2
certifi==2024.7.4
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.7
cloudscraper==1.2.71
colorama==0.4.6
cssselect==1.2.0
decorator==5.1.1
distlib==0.3.8
dnspython==2.6.1
EbookLib==0.18
email_validator==2.2.0
exceptiongroup==1.2.2
executing==2.0.1
fastapi==0.115.12
fastapi-cli==0.0.7
filelock==3.15.4
Flask==3.0.3
greenlet==3.1.1
grpcio==1.70.0
h11==0.16.0
html5lib==1.1
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
importlib_metadata==8.5.0
ipdb==0.13.13
ipython==8.12.3
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.6
jsonify==0.5
lxml==5.3.1
lxml_html_clean==0.4.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
mdurl==0.1.2
outcome==1.3.0.post0
packaging==24.2
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.4.0
pipenv==2024.0.1
platformdirs==4.2.2
prompt_toolkit==3.0.47
protobuf==5.29.3
ptyprocess==0.7.0
pure_eval==0.2.3
pycryptodome==3.21.0
pydantic==2.10.6
pydantic_core==2.27.2
pyease-grpc==1.7.0
PyExecJS==1.5.1
pygame==2.6.0
Pygments==2.18.0
pyparsing==3.1.4
PySocks==1.7.1
python-box==7.2.0
python-dotenv==1.0.1
python-multipart==0.0.20
python-slugify==8.0.4
PyYAML==6.0.2
questionary==2.1.0
readability-lxml==0.8.1
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rich==14.0.0
rich-toolkit==0.14.6
selenium==4.27.1
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
SQLAlchemy==2.0.41
stack-data==0.6.3
starlette==0.44.0
text-unidecode==1.3
tomli==2.0.1
tqdm==4.67.1
traitlets==5.14.3
trio==0.27.0
trio-websocket==0.11.1
typer==0.15.4
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.33.0
uvloop==0.21.0
virtualenv==20.26.3
watchfiles==0.24.0
wcwidth==0.2.13
webencodings==0.5.1
websocket-client==1.8.0
websockets==13.1
Werkzeug==3.0.0
wsproto==1.2.0
zipp==3.20.2
psycopg2-binary
fastapi==0.68.0
uvicorn==0.15.0
pydantic==1.8.2
sqlalchemy==1.4.25
python-dotenv==0.19.2
redis==4.3.4
fastapi-mail==0.3.12
bcrypt==3.2.0
python-jose[cryptography]==3.3.0
PyJWT==2.4.0
